Cherry-pick upstream r237073
Allow empty assembly string literal with -fno-gnu-inline-asm

Empty assembly string will not introduce assembly code in the output
binary and it is often used as a trick in the header to disable
optimizations. It doesn't conflict with the purpose of the option so it
is allowed with -fno-gnu-inline-asm flag.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@237073 91177308-0d34-0410-b5e6-96231b3b80d8

BUG= https://bugs.chromium.org/p/nativeclient/issues/detail?id=4333
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/1430233002 .